// src/views/aid/aimanage/types.ts
export interface ParamMapping {
  paramName: string;
  provider: string;
  providerParamName: string;
}